Distinguish renderless & Halogen components
This file change clarifies that the components linked are Halogen components specifically (and will not work with other frameworks). It also clarifies that some Halogen components are in the "renderless" style and are therefore used very differently than the all-in-one components. Finally, it adds a few more Halogen components from the ecosystem.