Typographically-correct apostrophes+quotation mark
While Pandoc and some Markdown renders can turn on typographic quotation
marks & apostrophes, not all do—including the renderer provided by
Microsoft GitHub. Due to the static nature of this document and where
it’s usually going to be read, rendered on Microsoft GitHub, I believe
it would be a better reading experience to use the correct symbols.

Changed the "hooks" → _hooks_ as double quotation marks, aka scare
quotes are for irony, and emphasis/italic is more appropriate.